Add v1.0.0 files
This may end up being the one and only release of this library (unless
issues are found that need to be fixed), as I have no plans to add more
features at the moment. The only exception might be more utility methods
for the Value enum, if I end up thinking of any more potentially useful
ones.